requirements.txt
# Django Core
Django>=5.0,<5.1
django-environ>=0.11.2
django-extensions>=3.2.3
django-allauth>=0.61.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-storages>=1.14.2
django-celery-beat>=2.5.0
django-htmx>=1.17.2

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Task Queue
celery>=5.3.6
redis>=5.0.1

# AI/ML
anthropic>=0.18.0

# PDF Processing
PyMuPDF>=1.23.8
pdfplumber>=0.10.3
Pillow>=10.2.0

# Payments
stripe>=8.0.0

# AWS S3 Storage
boto3>=1.34.0

# HTTP Client (for integrations)
httpx>=0.26.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
